+++ src/prng/index.js
@@ -1,0 +1,3 @@
+import prng from './prng';
+
+export default prng;
