+++ src/prng/select.js
@@ -1,0 +1,34 @@
+import {
+  ifElse,
+  isNil,
+  pipe,
+  propEq,
+  propSatisfies,
+  when,
+} from 'ramda';
+import seedrandom from 'seedrandom';
+
+const propNameEqualsArc4 = propEq('name', 'arc4');
+
+const setDefaultAlgorithm = ({ name, ...props }) => ({ ...props, algorithm: seedrandom });
+
+const propIsNil = propSatisfies(isNil);
+const whenPropIsNil = pipe(propIsNil, when);
+const whenSeedPropIsNil = whenPropIsNil('seed');
+const generateSeed = ({ ...props }) => ({ ...props, seed: seedrandom() });
+const generateSeedWhenPropIsNil = whenSeedPropIsNil(generateSeed);
+
+const setNamedAlgorithm = ({ name, ...props }) => ({ ...props, algorithm: seedrandom[name] });
+
+const generateSeedAndSetNamedAlgorithm = pipe(
+  generateSeedWhenPropIsNil,
+  setNamedAlgorithm,
+);
+
+const select = ifElse(
+  propNameEqualsArc4,
+  setDefaultAlgorithm,
+  generateSeedAndSetNamedAlgorithm,
+);
+
+export default select;
